// src/domain/models/duacoder.model.ts
export interface Duacoder {
  nif: string;                          
  name: string;                        
  biography?: string;                   
  department: string;                  
  position: string;                    
  skills: string[];                     
  photoUrl?: string;                   
  likesOnionTortilla: boolean;          
  birthDate?: Date;                     
}